split LoadShader to a new file.

+++ GLSLParse/GLSLTokenizer.cpp
@@ -543,7 +543,7 @@ bool GLSLTokenizer::IsKeyWord()
     {
         token_type = glsl_token_words[lastPossible].token;
         //token_length = strlen(glsl_token_words[lastPossible].word);
-        token_length = strlen(glsl_token_words[lastPossible].word);
+        token_length = (int) strlen(glsl_token_words[lastPossible].word);
         return true;
     }
